write tests for refactor code

+++ daos/campaign.go
@@ -44,7 +44,7 @@ func (dao *CampaignDAO) UpdateCampaign(input models.UpdateCampaignInput) (*model
 	return &campaign, nil
 }
 
-func (dao *CampaignDAO) DeleteCampaign(id uint64) (*models.Campaign, error) {
+func (dao *CampaignDAO) DeleteCampaign(id uint) (*models.Campaign, error) {
 	var campaign models.Campaign
 	if err := config.DB.Where("id = ?", id).First(&campaign).Error; err != nil {
 		return nil, err
